[feat] Clock description class with parser
This class (`ClkDescription`) contains the clock multi-tree for our soc.
Other operations will work on that tree. An important step of being able
to parse the serialization format also works already.

During parsing the system will validate the given data using the already
written json schema. This will make it easier to enforce the json
schema and keep it up-to-date.

Internally clocks are parsed into their specific classes and are
represented in a single self-referencing dictionary to represent the

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>