[feat]: Add graphing engine
The multi-tree can now be visualized using Graphviz Dot. It was a bit
difficult to get it to look somewhat reasonable as GraphViz has quite a
few problems. But it still seems to be the only tool out there for this
kind of job. (And I refuse to write my own plotting engine)

The Graph is read from left to right with the inputs and output all
being the same respective rank.

The ClkDescription also required extending as there are a few functions
over the graph, that are useful for visualization - but should not be
done by the Grapher subsystem (e.g. input and output clocks).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>