[feat]: Add filter ability with query filter
Using filters one can colorize or filter on the graph. Currently only
the QueryFilter is available, which filters on all inputs and outputs of
the given master node. How the filter should be displayed can be
configured in the Filter Accumulator. That is really basic rn, but the
idea is that it will be able to combine multiple filters and be
configurable from the CLI. Using it the grapher does not need to
understand or handle filters, except for the most basic way.

Filters are currently implemented on a lookup system, instead of them
extended the clk description / graph class. With that it is the
responsibility of the grapher to query the filters for every node and
edge on how it should be displayed.

With that filters can get really simple (e.g. only highlighting match
node) without the need to implement a complete graph, or arbitrarily
complex with even mirroring the graph state.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>