[feat]: SparseMemory class and Intel HEX parser
Add class to handle sparse memory representations and add corresponding
Intel HEX parser to it. Intel HEX is a memory data format that is
defined sparsely and as such quite commonly used to store memory dumps
or flash images for micro-controllers. It is not the most efficient one
as it is uncompressed and ASCII, but this is barely a problem on MCUs.

With this tool we will be able to implement the memory config filtering,
aka visualizing the current clock config of the MCU - the actual goal of
this project.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>