[feat] sparse mem: Implement segment cleanup method
The computational complexity of this classes getters and setters
increases linearly with the amount of segments in the internal storage.
As every call to __setitem__ increases the amount of segments by add
least one and parsing of e.g. Intel Hex uses __setitem__ for every
record, one can imagine how slow the class would become after parsing
just a single intel hex.

This is mostly a result of how I implemented __setitem__ with a slice
key, as I always insert the slice as a new segment and remove the parts
of the existing segments that would overlap this. This is significantly
easier to implement than trying to merge segments already here. As a
solution _cleanup_segments will now be called after every set, merging
continuous segments, cleaning our internal data structure.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>