[feat] Memory: MemoryClockGraph with parsing of memory
This required a few changes as it turned out that the previous state of
the program did not consider all cases.

- introduction of !addr32le and removal of !addr
When trying to access a certain register in memory, it is relevant to
know the width of it and its endianness when trying to parse said
register from a memory dump.

Previously the SOC config did not contain any such information. Instead
of declaring it globally I decided to do this register based, as there
are SOCs out there that mix register widths or even endianness.

In this process of experimentation I also added more fail-safes and
checks so that AddressObject will be initialized correctly.

- Add MemoryClockGraph
Currently this only does preliminary processing to test out whether
loading from memory actually works. In the future it will be able to be
used as a wrapper of ClockGraph to iterate on the graph build by the
config registers.

- SparseMemory: add get_register(self, addr)
Accessing the value of a register defined by an AddressObject is a bit
more involved, but always the same. So it makes sense to outsource this
to the SparseMemory module reducing the complexity at other points.

- Extending elements to have parser of memory
Instead of implementing these methods directly in MemoryClockGraph, I
decided to put the parse function more closely to where its specific
implementation is defined.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>