[feat]: Implement memory grapher
With this we can now finally visualize the clock map as described by the
configuration.

See clock-vis.py for an example setup. With that one can ask the graph
for a single clock overlaid over the complete graph. One can also give the mem graph
directly into the grapher to only visualize the clocks and edges that
are actually active.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>