[fix] filter: Allow returning list of properties
Previously I choose to only return a single property for each instance
of a filter, which during implementing a filter using said properties,
is a bit stupid. So here I extend it to lists of properties.

I also choose to have `lookup_clock_properties` to return a dict of
type[Property] -> Property so quick lookup is possible. This assumes
that every case where multiple properties could slash should be handled
explicitly (instead of giving that task to the grapher engine without
further thought).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>