[feat] filter: Add MemoryVisFilter
This filter is special in that it does not change the overall structure
of the graph, but instead injects properties derived from the
MemoryGraph into our filter subsystem. With this out of the way,
implementing visualization of the parsed memory into graph should be
easier.

This is also the reason why I choose to return a list of properties
instead of a single one as a clock can have a combination of the defined
memory properties (e.g. Registers + Mux Setting).

Specifically for the register overview (MemPropertyRegisters) a bit of
logic needed to be added to our elements.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>