[feat] grapher: Visualize properties
With this change most of what was needed from this tool is implemented,
as we can now visualize and display the memory section using our clock
graph.

For this to work I needed to switch from graphviz shape record to none /
plaintext and use html-like labels to describe the muxes instead of our
record ones. The reason is that record do not allow a change in
background color for a single cell, something very useful for indicating
which mux item is selected.

The change also adds a dashed border to disabled clocks and adds the
current registers involved in a clock state to the label so one can
easily lookup how a clock is configured in detail by taking a look into
the data sheet (specifically relevant for divisors).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>