[feat] filter: accumulator: Allow hiding of hidden
Previous iterations simply hard-coded the color table of the
FilterAccumulator. As a feature of the system is that we can color or
even hide elements having the specified state, it is only sensible to
allow exposing this to the outside.

This is a bit of a short-term fix for a more complete system later on.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>