[fix] SparseMemory: __get__(slice) retrieved wrong elements
I made a mistake in the slicing of the segment for the branch processing
the case that our overlapping segment was left of the slice.

In reality this would not have changed anything in the result, as python
allows negative indexing and that formula would result in
  -(len(seg) - (key.start - seg.start))

It can still be considered a fix, as this was not the intended way for
the expression to be understood - even if it worked out in the end.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>