Add pinned requirements file
Pythons package environment is a bit stupid I dare say. Using pip freeze
is encouraged (for good reasons - e.g. API changes), but it also means
that one cannot ever really update the requirements easily. Because what
is now required by a package installed with pip and which is actually
needed by the tool.

So yeah. I fan of providing a lock file for "known good state" and a
requirements file that does not pin anything so that during development
one can actually update the dependencies (and check for new failures).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>