[fix] Grapher: set subgraph font size to 14 again
The fontsize of graphs (so their labels and such) is set to 40 so that
the overall graph title / label can be displayed prominently. What I did
not consider is that this also changes the font size for all subgraphs.
Graphviz default for fontsize is 14, so setting it back to that.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>